Add initial set of (old|un)stable Dockerfile files
Using official "golang" image as our base.

- oldstable
  - the still supported, but outgoing Go version
  - 1.13.14
- unstable
  - the upcoming, not officially supported Go version
  - 1.15rc1
- stable
  - the current production Go version
  - 1.14.6

A Makefile is included to help drive the building and
uploading process.

These Docker files lock-in specific dependency versions:

- staticcheck
  - 2020.1.5
- golangci-lint
  - v1.29.0
- bsdmainutils
  - 11.1.2+b1
- tree
  - 1.8.0-1

The latter two are either used with current Makefile-based
project builds or (in the case of `tree`) considered for use.

Using Docker Hub to storage images as the GitHub Package Repos
require authentication for public package downloads and this
access is ... problematic for use between multiple projects for
CI use.