Fix linting-only image environment
Unfortunately the last release used an Alpine Linux
base image that was lacking required support for
the `net` package. I did not dig deep to discover
the most optimal fix, but the workaround employed
here is to switch the base image to the same Docker
image provided by the golangci-lint project
and just drop-in the staticcheck binary to complete
our package.

While this does substantially increase the image
size, it still weighs in a little less than the
shared image used for other purposes.

This can be further researched at a later time.