Bundle golangci-lint, markdownlint config files
By including both of these config files at the root of the
generated containers we provide a usable linting configuration
baseline for all Go projects using these containers.

NOTE: As of this commit, markdownlint is not included in the
containers generated by this project. This linter should be
installed within an appropriate environment (e.g., via
GitHub Actions) and then directed to use the
`.markdownlint.yml` file provided by this container.

refs GH-32