Dependabot: Move Dockerfiles into separate subdirs
Instead of grouping all stable Dockerfiles together,
separate into subdirectories to provide Dependabot
an isolated view of the specific Docker image that
we wish for it to help maintain.

Update the Makefile to reflect the new paths.

Update the GitHub Actions Workflow for Dockerfile
linting to reflect the new paths, including some
Dockerfiles that I previously failed to include in
the list to check.

refs GH-204