Remove explicit Go version used by golangci-lint
If not explicitly not, golangci-lint will use the
available go.mod file to determine what Go features
and applicable linting rules to use.

For example, if the project specifies Go 1.14 as the target
Go version, golangci-lint will not issue a warning if the
io/ioutil package is used (deprecated in Go 1.16).

refs GH-716