Update Go 1.20.0 images to 1.20.1
When I last updated the stable & unstable Dockerfiles to
use the latest Go 1.20 images I use the standard Go version
convention of "1.20" instead of "1.20.0". The official Go
Docker images use 1.20 as a sort of "floating" tag to reference
the latest version from the 1.20 series and not explicitly
1.20.0.

This commit explicitly references the latest 1.20.1 version.

refs af1b7eff24278f8ebe8b41bce2861893042d96a1
refs https://github.com/dependabot/dependabot-core/issues/6672