Rebrand, update `go-ci-stable-debian-build` image
The `go-ci-stable-debian-build` image was poorly named and
failed to convey the intended purpose:

Provide an image to support cross-platform, static cgo-enabled
builds for Windows and Linux.

This set of changes reworks the image to drop linting support
(which was always an afterthought) and rename it to
`go-ci-stable-cgo-mingw-w64-build`.

The README has been updated to remove the "Do you use this
image?" prompt and provide an overview that better reflects
the image's purpose.

NOTE: The previous steps to create a temporary copy of the
golangci-lint config file for each stable image being built
and then prune the temporary file as part of the dedicated
cleanup recipe and post-build cleanup have been removed.
The "stable" variant of the linting config file has been
moved directly into the stable/combined/ path.

refs GH-657