Process all images when building
When running the `build` Makefile recipe build legacy mirror
images in addition to all other images. This adds to the build
time, but helps ensure that all images are uploaded and that the
(newly created) `go-ci-oldstable-mirror-build` and
`go-ci-stable-mirror-build` image tags are maintained as
intended.

If all goes as planned then the legacy images will grow
additional versioned tags over time, but the layers for the
images will *not* continue to grow (as the upstream images
are unmaintained, but still provided as-is).

Drop separate workflow job step to handle building legacy
images now that the `build` recipe also triggers legacy
image builds.

refs GH-847