Update Dockerfiles to specify chosen base image OS
These images were updated to use the most recent stable
base image OS available:

- mirror images
- oldstable images
- stable images

The bump to Go 1.21rc2 for the unstable images also included
the change to specify the base image OS.

NOTE:

Going forward the base image OS will likely stay the same for
the life of the specific series reflected in the image type.

For example, Debian "bookworm" was recently released. When
Go 1.21.0 is released as the next stable version, the "stable"
Debian-based images will continue to use the "bookworm" OS
as the base image for the life of the Go 1.21 series. If a
newer Debian OS version is available when Go 1.22.0 is released
then the "stable" images will adopt that Debian OS as the base
image OS for Debian-based images. The 1.21 series will continue
to use the former Debian OS as the base image OS (for
Debian-based images).

refs GH-1057
refs GH-1058