Replace choffmeister/git-describe-semver GH action
Replace the (currently) problematic GitHub Action used in the scheduled
monthly image build workflow with a custom job (`git_describe_semver`)
that explicitly uses the v0.3.11 version of the Docker container to
generate the release asset release version.

That version is captured as the job's `version` output which is
referenced by the `build_all_images_using_makefile` job in the
scheduled-monthly workflow. That job is set as dependent on the
`git_describe_semver` job so that they only run after the asset
release version is computed.

Since the value is computed once at the start of the workflow,
I've updated the `actions/checkout` step for the
`build_all_images_using_makefile` job to use the default checkout
option(s) instead of pulling the full repo history each time. This
should (somewhat) optimize that step in the build process.

refs GH-1299
refs GH-1276