Update oldstable Go version for tools/go.mod file
One of the dependencies we're tracking has bumped their minimum
Go version, which caused Dependabot tracked dependency updates
to fail when processing the `tools/go.mod` file. This surfaced
the fact that the Go version in the file hasn't been bumped in
a while.

Updating intent from tracking the very latest stable series to
tracking the oldstable series instead (with the assumption that
it will be "new enough" for deps tracking).