Fix unstable image baseimage references
Replace `1.23` (series) references with `1.23.0`
(explicit version) so that image rebuilds can continue to
work as intended by providing a very specific Go version (and
not just the latest in the series when the image happens to be
built).

This was the intended outcome with the recent 1.23rc2 baseimage
reference updates, though I missed this omission when merging
those PRs.