Update unstable build images to Go 1.25rc1
The unstable build images have been updated to use the
Go 1.25rc1 base image while the unstable "combined" image
continues to use the `amd64/golang:1.24.4-bookworm` base
image for now.

Once the staticcheck and golangci-lint tools are updated to
support Go 1.25 we can switch the base image for the "combined"
linter image to the Go 1.25 series.

refs GH-2074