WORKROUND: Pin nfpm to v2.43.1 in oldstable images
Workaround current incompatibility with nfpm v2.43.2 (and newer)
and "oldstable" Go releases from the 1.24 series by pinning
the nfpm version to v2.43.1.

Future plans to handle this include building nfpm with the
latest version of Go and bundling within the oldstable image
instead of attempting to build using the oldstable release
series as we're currently doing.